feat: adjust ID of search input
Firefox adds recently used search terms to the suggestion list. The
problem is that, it grabs the terms from every `<input>` element with a
same element ID of arbitrary website, so using a generic ID like
`search` will make the suggestion list populated with search terms from
random websites. But it is not very likely that you want to search our
calendar for a same term you used in Google, for instance. Changing the
ID should reduce the possibility of such a collision.