fix(Duration): remove day componet from `toDatetime`
Although `1D` is interpreted as 86400 seconds in HTML Living Standard,
it is defined as a calendar day in ISO 8601, which may vary depending on
daylight saving time or leap seconds. Since both of HTML Living Standard
and ISO 8601 do not prohibit the hour component from exceeding 24, let's
remove the day component to avoid the potential disambiguity.