fix(a11y): remove `aria-live` from `<main>`
The event duration view of ongoing live streams are updated every
second, causing the event description to be read by screen readers every
second! Setting `aria-live="on"` didn't work, so I removed the
`aria-live="polite"` entirely, as the main view is almost stable once
all the feeds are loaded, until which the region is marked as busy
anyway.