refactor: put `model.search` and `model.feeds` in single record
We have to put at least a pair of arguments to `viewMain` in a single
record because there is no such thing as `Html.Lazy.lazy9`. Previously,
we were putting `now` and `timeZone` together, but `search` and `feeds`
seemed to be more often used together so I put them together instead.