refactor: flatten module structure
I'm not a big fan of mixing namespace of modules, but this seems to be
the idiomatic way in Elm so let's follow the convention.

I'm also not satisfied with the fact that you cannot list local module
imports and remote ones separately when formatting, but maybe that
should be kept for an elm-format-feature-request wishlist or something.