feat(a11y): make labels more explicit
Previously, we were hiding visual elements if they were labelled by
other elements, in which case the elements _seemed_ to be irrelevant to
screen reader users. However, screen reader users may use pointing
device for quick navigation inside the display, e.g. with VoiceOver's
Trackpad Commander mode. In such a situation, the `aria-hidden="true"`
elements would look like a void in the display for screen reader users
and could not be interacted with the pointer.