feat(i18n): use `Intl.RelativeTimeFormat`
`due` translation now displays something like `due 1 小時前` in
unsupported locale. I think that it is acceptable given that the user is
already experiencing mixed locale in other parts of the app, and that
the benefit of localizing more places outweighs even if it is partial
(as long as proper metadata like `lang` are provided).