feat: sort search suggestions by "trendiness"
Previously, we were naively sorting the terms by number of occurences,
which would result in less informative terms like "#けもV" being on the
top.

The new approach is not perfect though. For example, terms appeared in
the previous week but not in the current week are put under terms that
appeared equally in both weeks, which includes common terms like
"#けもV", but it is not obvious that the behavior is desirable.

Another example is that "trendiness" as in the current definition is
somewhat unstable: trendiness of a term that appeared 6 days ago will
slump in the next day. Maybe we should employ some easing measure like
moving average? I don't know.

Also, the time window of 7 days might become insufficient when KemoV
gets bigger, or, in a shorter term, when we start to collect collab
streams hosted by extra-KemoV channels, where the first page of the
feed may be unable to contain the whole two-week period.

Anyway, those are left for future improvements. Current approach puts
terms like "#けもV" at seemingly moderate positions at least, which is
enough for the time being.