refactor(a11y): add `lang` to status text (or preparation thereof)
Unfortunately, VoiceOver on Safari and Chromium still does not respect
the `lang` attribute when reading the live region (and Firefox does not
seem to respect _any_ `lang` attribute to begin with).

Having said that, separating the rendering of the status text into its
own module seems to be a reasonable refactoring which should make future
accessibility feature additions easier. It's still a bit unfortunate
that we now have an off-screen element instead of an `aria-label`, but
I think it's an acceptable change considering `aria-label`'s lack of
semantic structures.