chore(feeds): update icon URLs
We are now using images from the official website, as images from
YouTube break once in a while.

Because how closely the images on the official site resemble the icons
from the corresponding YouTube channel varies from page to page, we are
using different source pages on an ad-hoc basis.